Remove Notification realted infrastructure again. Moving to SpeziNotifications
