+++ Sources/Spezi/Spezi.docc/Interactions with Application.md
@@ -15,7 +15,7 @@ SPDX-License-Identifier: MIT
 ## Overview
 
 Spezi provides platform-agnostic mechanisms to interact with your application instance.
-To access application properties or actions you can use the ``Application`` property wrapper within your
+To access application properties or actions you can use the ``Module/Application`` property wrapper within your
 ``Module``, ``Standard`` or SwiftUI `View`.
 
 > Tip: The <doc:Notifications> articles illustrates how you can easily manage user notifications within your Spezi application. 
@@ -24,7 +24,7 @@ To access application properties or actions you can use the ``Application`` prop
 
 ### Application Interaction
 
-- ``Application``
+- ``Module/Application``
 
 ### Properties
 
@@ -35,8 +35,6 @@ To access application properties or actions you can use the ``Application`` prop
 
 - ``Spezi/registerRemoteNotifications``
 - ``Spezi/unregisterRemoteNotifications``
-- ``Spezi/notificationSettings``
-- ``Spezi/requestNotificationAuthorization``
 
 ### Platform-agnostic type-aliases
