+++ Sources/Spezi/Spezi/EnvironmentValues+Spezi.swift
@@ -13,8 +13,11 @@ extension EnvironmentValues {
     private struct SpeziKey: EnvironmentKey {
         static let defaultValue: Spezi? = nil
     }
-
-    var spezi: Spezi? {
+    
+    /// Access the Spezi instance to provide additional support for the SwiftUI environment.
+    ///
+    /// Use this property as a basis for your own
+    var spezi: Spezi? { // TODO: we do not really need this?
         get {
             self[SpeziKey.self]
         }
