+++ Sources/Spezi/Utilities/Application+TypeAlias.swift
@@ -26,7 +26,9 @@ public typealias _Application = NSApplication // swiftlint:disable:this type_nam
 public typealias _Application = WKApplication // swiftlint:disable:this type_name
 
 extension WKApplication {
-    static var shared: WKApplication {
+    /// Allow the same access pattern for WKApplication. Bridges to the `shared()` method.
+    @_documentation(visibility: internal)
+    public static var shared: WKApplication {
         shared()
     }
 }
