+++ Tests/UITests/TestApp/TestApp.swift
@@ -10,6 +10,8 @@ import Spezi
 import SwiftUI
 import XCTestApp
 
+// TODO: test @Application somewhere! (maybe just use @Environment???)
+
 
 @main
 struct UITestsApp: App {
